// src/app/model/shared/arquivos.ts
export class Arquivos {
  nomeImage?: string;
  // tslint:disable-next-line:variable-name
  nome_img?: string;
  pathImage?: string;
  tipo?: string;
  // tslint:disable-next-line:variable-name
  tipo_foto?: string;
  // tslint:disable-next-line:variable-name
  pasta_arquivo: string;
  imagem: any;
  file: any;
  formato: string;
  // tslint:disable-next-line:variable-name
  momento_fotos: string;
  // tslint:disable-next-line:variable-name
  id_pedido: any;
  // tslint:disable-next-line:variable-name
  id_cliente: any;
  cliente: any;
  // tslint:disable-next-line:variable-name
  id_locacao: any;
  descricao: string;

}